Android 11 device controls - support dimmer & error logging
