+++ mobile/src/main/AndroidManifest.xml
@@ -233,7 +233,7 @@
             tools:targetApi="n"
             tools:ignore="Instantiatable" />
 
-        <service android:name=".background.ControlsService"
+        <service android:name=".background.ItemsControlsProviderService"
                  android:label="@string/app_name"
                  android:permission="android.permission.BIND_CONTROLS"
                  tools:targetApi="R">
